Moved static methods out of ``ResourceViewSet``
* To new module ``utils.py``

* Cleaned up some stale imports in ``views.py``
